Tidy-up web-component JavaScript files
