+++ django_app/frontend/src/js/citations.js
@@ -1,0 +1,1 @@
+import "./web-components/markdown-converter.js";
