+++ django_app/frontend/src/js/main.js
@@ -1,0 +1,3 @@
+import "./trusted-types.js";
+import "./libs/govuk-frontend.min.js";
+import "../../node_modules/i.ai-design-system/dist/iai-design-system.js";
