+++ django_app/frontend/src/js/web-components/chats/chat-controller.js
@@ -1,0 +1,56 @@
+// @ts-check
+
+class ChatController extends HTMLElement {
+  connectedCallback() {
+    const messageForm = this.closest("form");
+    const textArea = /** @type {HTMLInputElement | null} */ (
+      this.querySelector(".js-user-text")
+    );
+    const messageContainer = this.querySelector(".js-message-container");
+    const insertPosition = this.querySelector(".js-response-feedback");
+    const feedbackButtons = /** @type {HTMLElement | null} */ (
+      this.querySelector("feedback-buttons")
+    );
+    let selectedDocuments = [];
+
+    messageForm?.addEventListener("submit", (evt) => {
+      evt.preventDefault();
+      const userText = textArea?.value.trim();
+      if (!textArea || !userText) {
+        return;
+      }
+
+      let userMessage = document.createElement("chat-message");
+      userMessage.setAttribute("data-text", userText);
+      userMessage.setAttribute("data-role", "user");
+      messageContainer?.insertBefore(userMessage, insertPosition);
+
+      let aiMessage = /** @type {ChatMessage} */ (
+        document.createElement("chat-message")
+      );
+      aiMessage.setAttribute("data-role", "ai");
+      messageContainer?.insertBefore(aiMessage, insertPosition);
+      aiMessage.stream(
+        userText,
+        selectedDocuments,
+        this.dataset.sessionId,
+        this.dataset.streamUrl || "",
+        this
+      );
+      /** @type {HTMLElement | null} */ (
+        aiMessage.querySelector(".iai-chat-bubble")
+      )?.focus();
+
+      // reset UI
+      if (feedbackButtons) {
+        feedbackButtons.dataset.status = "";
+      }
+      textArea.value = "";
+    });
+
+    document.body.addEventListener("selected-docs-change", (evt) => {
+      selectedDocuments = /** @type{CustomEvent} */ (evt).detail;
+    });
+  }
+}
+customElements.define("chat-controller", ChatController);
