+++ django_app/frontend/src/js/web-components/chats/chat-message.js
@@ -1,59 +1,5 @@
 // @ts-check
 
-class SourcesList extends HTMLElement {
-  constructor() {
-    super();
-    this.sources = [];
-  }
-
-  /**
-   * Adds a source to the current message
-   * @param {string} fileName
-   * @param {string} url
-   */
-  add = (fileName, url) => {
-    this.sources.push({
-      fileName: fileName,
-      url: url,
-    });
-
-    let html = `
-            <h3 class="iai-chat-bubble__sources-heading govuk-heading-s govuk-!-margin-bottom-1">Sources</h3>
-            <div class="iai-display-flex-from-desktop">
-            <ul class="govuk-list govuk-list--bullet govuk-!-margin-bottom-0">
-        `;
-    this.sources.forEach((source) => {
-      html += `
-                <li class="govuk-!-margin-bottom-0">
-                    <a class="iai-chat-bubbles__sources-link govuk-link" href="${source.url}">${source.fileName}</a>
-                </li>
-            `;
-    });
-    html += `</div></ul>`;
-
-    this.innerHTML = html;
-  };
-
-  /**
-   * Shows to citations link/button
-   * @param {string} chatId
-   */
-  showCitations = (chatId) => {
-    if (!chatId) {
-      return;
-    }
-    let link = document.createElement("a");
-    link.classList.add(
-      "iai-chat-bubble__citations-button",
-      "govuk-!-margin-left-2"
-    );
-    link.setAttribute("href", `/citations/${chatId}`);
-    link.textContent = "View information behind this answer";
-    this.querySelector(".iai-display-flex-from-desktop")?.appendChild(link);
-  };
-}
-customElements.define("sources-list", SourcesList);
-
 class ChatMessage extends HTMLElement {
   connectedCallback() {
     const uuid = crypto.randomUUID();
@@ -229,94 +175,3 @@ class ChatMessage extends HTMLElement {
   };
 }
 customElements.define("chat-message", ChatMessage);
-
-class ChatController extends HTMLElement {
-  connectedCallback() {
-    const messageForm = this.closest("form");
-    const textArea = /** @type {HTMLInputElement | null} */ (
-      this.querySelector(".js-user-text")
-    );
-    const messageContainer = this.querySelector(".js-message-container");
-    const insertPosition = this.querySelector(".js-response-feedback");
-    const feedbackButtons = /** @type {HTMLElement | null} */ (
-      this.querySelector("feedback-buttons")
-    );
-    let selectedDocuments = [];
-
-    messageForm?.addEventListener("submit", (evt) => {
-      evt.preventDefault();
-      const userText = textArea?.value.trim();
-      if (!textArea || !userText) {
-        return;
-      }
-
-      let userMessage = document.createElement("chat-message");
-      userMessage.setAttribute("data-text", userText);
-      userMessage.setAttribute("data-role", "user");
-      messageContainer?.insertBefore(userMessage, insertPosition);
-
-      let aiMessage = /** @type {ChatMessage} */ (
-        document.createElement("chat-message")
-      );
-      aiMessage.setAttribute("data-role", "ai");
-      messageContainer?.insertBefore(aiMessage, insertPosition);
-      aiMessage.stream(
-        userText,
-        selectedDocuments,
-        this.dataset.sessionId,
-        this.dataset.streamUrl || "",
-        this
-      );
-      /** @type {HTMLElement | null} */ (
-        aiMessage.querySelector(".iai-chat-bubble")
-      )?.focus();
-
-      // reset UI
-      if (feedbackButtons) {
-        feedbackButtons.dataset.status = "";
-      }
-      textArea.value = "";
-    });
-
-    document.body.addEventListener("selected-docs-change", (evt) => {
-      selectedDocuments = /** @type{CustomEvent} */ (evt).detail;
-    });
-  }
-}
-customElements.define("chat-controller", ChatController);
-
-class DocumentSelector extends HTMLElement {
-  connectedCallback() {
-    const documents = /** @type {NodeListOf<HTMLInputElement>} */ (
-      this.querySelectorAll('input[type="checkbox"]')
-    );
-
-    const getSelectedDocuments = () => {
-      let selectedDocuments = [];
-      documents.forEach((document) => {
-        if (document.checked) {
-          selectedDocuments.push(document.value);
-        }
-      });
-      const evt = new CustomEvent("selected-docs-change", {
-        detail: selectedDocuments,
-      });
-      document.body.dispatchEvent(evt);
-    };
-
-    // update on page load
-    getSelectedDocuments();
-
-    // update on any selection change
-    documents.forEach((document) => {
-      document.addEventListener("change", getSelectedDocuments);
-    });
-  }
-}
-customElements.define("document-selector", DocumentSelector);
-
-// Update URL when a new chat is created
-document.addEventListener("chat-response-end", (evt) => {
-  const sessionId = /** @type{CustomEvent} */ (evt).detail.session_id;
-  window.history.pushState({}, "", `/chats/${sessionId}`);
-});
