+++ django_app/frontend/src/js/web-components/chats/document-selector.js
@@ -1,0 +1,31 @@
+// @ts-check
+
+class DocumentSelector extends HTMLElement {
+  connectedCallback() {
+    const documents = /** @type {NodeListOf<HTMLInputElement>} */ (
+      this.querySelectorAll('input[type="checkbox"]')
+    );
+
+    const getSelectedDocuments = () => {
+      let selectedDocuments = [];
+      documents.forEach((document) => {
+        if (document.checked) {
+          selectedDocuments.push(document.value);
+        }
+      });
+      const evt = new CustomEvent("selected-docs-change", {
+        detail: selectedDocuments,
+      });
+      document.body.dispatchEvent(evt);
+    };
+
+    // update on page load
+    getSelectedDocuments();
+
+    // update on any selection change
+    documents.forEach((document) => {
+      document.addEventListener("change", getSelectedDocuments);
+    });
+  }
+}
+customElements.define("document-selector", DocumentSelector);
