+++ django_app/redbox_app/templates/chats.html
@@ -155,12 +155,7 @@
 
 </div>
 
-
 <script src="{{ static('js/libs/showdown.min.js') }}"></script>
-<script src="{{ static('js/chats/markdown.js') }}"></script>
-<script src="{{ static('js/chats/feedback.js') }}"></script>
-<script src="{{ static('js/chats/message-input.js') }}"></script>
-<script src="{{ static('js/chats/streaming.js') }}"></script>
-<script src="{{ static('js/chats/chat-title.js') }}"></script>
+<script type="module" src="{{ static('js/chats.js') }}"></script>
 
 {% endblock %}
