+++ django_app/redbox_app/templates/citations.html
@@ -19,6 +19,6 @@
 </div>
 
 <script src="{{ static('js/libs/showdown.min.js') }}"></script>
-<script src="{{ static('js/chats/markdown.js') }}"></script>
+<script src="{{ static('js/citations.js') }}"></script>
 
 {% endblock %}
